Create test_anvil2.css
Hey 👋 Proposing this new file addition as I've been working to try and make our CSS migration a two-step process. We need a quick MVP that we can iterate on without updating the legacy stylesheet. The long term fix is being proposed by an ADR and will be implemented when we have some downtime 🤞